+++ core/src/Logger.kt
@@ -3,7 +3,8 @@ package klite
 
 import java.lang.System.Logger.Level.*
 
-fun Any.logger(name: String = javaClass.name): System.Logger = System.getLogger(name)
+fun logger(name: String): System.Logger = System.getLogger(name)
+fun Any.logger(): System.Logger = System.getLogger(javaClass.name)
 inline fun System.Logger.debug(msg: String) = log(DEBUG, msg)
 inline fun System.Logger.info(msg: String) = log(INFO, msg)
 inline fun System.Logger.warn(msg: String) = log(WARNING, msg)