+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 1.6.9
 * core: logger(name) is now accessible without a class instance context
 * core: Any.logger() will now take the closest non-anonymous superclass
 * server: errors.on<SomeException>(StatusCode) convenience reified function added